Separate models for mongo and sql
